Fix more screen testable
