+++ app/src/main/AndroidManifest.xml
@@ -5,6 +5,7 @@
     <uses-permission android:name="android.permission.INTERNET" />
 
     <application
+        android:name=".MyApplication"
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
         android:fullBackupContent="@xml/backup_rules"
@@ -14,7 +15,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.AndroidProject2025.compose">
         <activity
-            android:name=".ui.main.MainActivity"
+            android:name=".ui.feature.main.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.AndroidProject2025.compose">
@@ -26,7 +27,7 @@
         </activity>
 
         <activity
-            android:name=".ui.detail.DetailActivity"
+            android:name=".ui.feature.detail.DetailActivity"
             android:exported="false"
             android:theme="@style/Theme.AndroidProject2025.xml" />
     </application>
