+++ app/src/main/java/com/sample/android/MyApplication.kt
@@ -1,0 +1,7 @@
+package com.sample.android
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class MyApplication : Application()
