+++ app/src/main/java/com/sample/android/ui/feature/detail/model/DetailUiData.kt
@@ -1,0 +1,7 @@
+package com.sample.android.ui.feature.detail.model
+
+class DetailExtraData {
+    companion object {
+        const val KEY_DETAIL_ACTIVITY_RESULT = "key_detail_activity_result"
+    }
+}
