+++ app/src/main/java/com/sample/android/ui/feature/main/model/UserUiData.kt
@@ -1,4 +1,4 @@
-package com.sample.android.ui.data
+package com.sample.android.ui.feature.main.model
 
 import android.os.Parcelable
 import androidx.annotation.Keep
