+++ app/src/test/java/com/sample/android/ui/data/SearchTabDataTest.kt
@@ -1,6 +1,12 @@
 package com.sample.android.ui.data
 
 import com.sample.android.data.UserMetaData
+import com.sample.android.ui.feature.main.model.SearchTabBorder
+import com.sample.android.ui.feature.main.model.SearchTabData
+import com.sample.android.ui.feature.main.model.SearchTabMetaData
+import com.sample.android.ui.feature.main.model.UserUiData
+import com.sample.android.ui.feature.main.model.like
+import com.sample.android.ui.feature.main.model.unlike
 import junit.framework.TestCase.assertEquals
 import junit.framework.TestCase.assertFalse
 import junit.framework.TestCase.assertTrue
